+++ icinga2/resource_icinga2_host_test.go
@@ -99,27 +99,28 @@ func TestAccCreateVariableHost(t *testing.T) {
 	})
 }
 
-func TestAccCreateTemplateHost(t *testing.T) {
-	testAccCreateTemplateHost := `resource "icinga2_host" "tf-4" {
-	hostname = "terraform-host-4"
-	address = "10.10.10.4"
-	check_command = "hostalive"
-	templates = ["generic", "az1"]
-}`
+func TestAccCreateZoneHost(t *testing.T) {
+
+	var testAccCreateVariableHost = fmt.Sprintf(`
+		resource "icinga2_host" "tf-4" {
+		hostname = "terraform-host-4"
+		address = "10.10.10.4"
+		check_command = "hostalive"
+		zone = "master"
+		}`)
+
 	resource.Test(t, resource.TestCase{
 		PreCheck:  func() { testAccPreCheck(t) },
 		Providers: testAccProviders,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccCreateTemplateHost,
+				Config: testAccCreateVariableHost,
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckHostExists("icinga2_host.tf-4"),
 					testAccCheckResourceState("icinga2_host.tf-4", "hostname", "terraform-host-4"),
 					testAccCheckResourceState("icinga2_host.tf-4", "address", "10.10.10.4"),
 					testAccCheckResourceState("icinga2_host.tf-4", "check_command", "hostalive"),
-					testAccCheckResourceState("icinga2_host.tf-4", "templates.#", "2"),
-					testAccCheckResourceState("icinga2_host.tf-4", "templates.0", "generic"),
-					testAccCheckResourceState("icinga2_host.tf-4", "templates.1", "az1"),
+					testAccCheckResourceState("icinga2_host.tf-4", "zone", "master"),
 				),
 			},
 		},